% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crypto_currencies.R
\name{get_token_id}
\alias{get_token_id}
\title{Get Token Id}
\usage{
get_token_id(symbol)
}
\arguments{
\item{symbol}{The cryptocurrency symbol in lowercase}
}
\value{
The cryptocurrency ID
}
\description{
This function can be used to get an id which can be used in get_token_price() method.
}
\examples{
btcID = get_token_id("btc");
bnbID = get_token_id("bnb");
}
