% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crypto_currencies.R
\name{get_token_price}
\alias{get_token_price}
\title{Get Token Price}
\usage{
get_token_price(id, days, interval)
}
\arguments{
\item{id}{The cryptocurrency id, get it with get_token_id() method}

\item{days}{The days in the past since you want to get the data.}

\item{interval}{"daily" or "hourly", note that chosing "hourly" the max days returned are 90}
}
\value{
A tibble containing the cryptocurrency history
}
\description{
This function can be used to get the price in time of a cryptocurrency.
}
\examples{
btc = get_token_price("bitcoin");
bnb = get_token_price("binance-coin");
}
